requirements.txt
# ML & Data
scikit-learn==1.8.0
pandas==2.3.3
numpy==2.4.2
joblib==1.5.3
imbalanced-learn==0.14.1
xgboost==3.1.3

# Streamlit (Deploy Principal - Datathon)
streamlit==1.54.0
plotly==5.24.1
matplotlib==3.10.8
seaborn==0.13.2

# API (Complementar)
fastapi==0.128.6
uvicorn[standard]==0.40.0
pydantic==2.12.5
python-multipart==0.0.22

# Database
supabase==2.27.3

# Monitoring
evidently==0.7.20

# HuggingFace
huggingface-hub==1.4.1

# Testing
pytest==9.0.2
pytest-cov==7.0.0
pytest-asyncio==1.3.0
httpx==0.28.1

# Utilities
python-dotenv==1.2.1
loguru==0.7.3
